refactor: Moved test case to App.test.js
