+++ src/App.test.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import App from './App';
 
 test('Succesfully renders the submit button', () => {
@@ -20,4 +21,15 @@ test('selecting an option changes the paragraph text', () => {
 
   const selectedText = screen.getByText('You selected Mediterranean Diet');
   expect(selectedText).toBeInTheDocument();
+});
+
+test('food item is searched when button is clicked', () => {
+  render(<App />);
+  const button = screen.getByText('Search');
+  fireEvent.click(button);
+  const input = screen.getByPlaceholderText('Search...');
+
+  userEvent.type(input, 'chicken');
+  expect(screen.getByPlaceholderText('Search...')).toHaveValue('chicken');
+  userEvent.type(input);
 });
